Function ncollide_procedural::unit_quad

pub fn unit_quad<N, P, V>(usubdivs: usize, vsubdivs: usize) -> TriMesh<N, P, V> where N: Scalar, P: Point<N, V>, V: Vect<N>

Adds a double-sided quad with unit size to the scene.

The quad is initially centered at (0, 0, 0). Its normal is the z axis. The quad itself is composed of a user-defined number of triangles regularly spaced on a grid. This is the main way to draw height maps.

Arguments

usubdivs - number of horizontal subdivisions. This correspond to the number of squares which will be placed horizontally on each line. Must not be 0.

vsubdivs - number of vertical subdivisions. This correspond to the number of squares which will be placed vertically on each line. Must not be 0.