Module ncollide_procedural::path

Path generation.

Structs

ArrowheadCap
A cap that looks like an arrow.

NoCap
A cap that renders nothing.

PolylinePath
A path with its sample points given by a polyline.

PolylinePattern
A pattern composed of polyline and two caps.

Enums

PathSample
A sample point and its associated tangent.

Traits

CurveSampler
A curve sampler.

PolylineCompatibleCap
Trait to be implemented by caps compatible with a PolylinePattern.

StrokePattern
A pattern that is replicated along a path.