Struct ncollide_procedural::path::ArrowheadCap

pub struct ArrowheadCap<N> {
    // some fields omitted
}

A cap that looks like an arrow.

Methods

impl<N> ArrowheadCap<N> where N: Scalar

fn new(radius_scale: N, front_dist_to_head: N, back_dist_to_head: N) -> ArrowheadCap<N>

Creates a cap that looks like an arrow.

Arguments:

radius_scale - scale factor of the cap base.

front_dist_to_head - distance from the path endpoint and the arrow tip.

back_dist_to_head - distance from the path endpoint and the cap base.